shows active Block in Topic View
