+++ static/entry.js
@@ -1,5 +1,5 @@
-window.Grid = require('babel!./grid.js');
-window.Timetable = require('babel!./personalTimetable.js').default;
+window.Grid = require('babel!./views/grid.js');
+window.Timetable = require('babel!./views/personalTimetable.js');
 window.React = require('react');
 
 require('Openspace.Ui.Stream').main();
